intelligent_investor_India/config/settings.py
# config/settings.py
import os
from pathlib import Path
from dotenv import load_dotenv

load_dotenv()

BASE_DIR = Path(__file__).resolve().parent.parent 
DATA_DIR = BASE_DIR / "data"
REPORTS_DIR = BASE_DIR / "reports"

DATA_DIR.mkdir(parents=True, exist_ok=True)
REPORTS_DIR.mkdir(parents=True, exist_ok=True)

# --- INDIAN SETTINGS ---
STARTING_CAPITAL = 100000  # ₹1,00,000 (Example Capital)
CURRENCY_SYMBOL = "₹"