intelligent_investor_India/requirements.txt
streamlit
pandas
yfinance
plotly
numpy
python-dotenv